after fix

+++ L2CacheTest/Migrations/TestContextModelSnapshot.cs
@@ -20,9 +20,9 @@ namespace L2CacheTest.Migrations
 
             modelBuilder.Entity("L2CacheTest.Entry", b =>
                 {
-                    b.Property<long>("Id")
+                    b.Property<int>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("bigint")
+                        .HasColumnType("int")
                         .HasAnnotation("SqlServer:ValueGenerationStrategy", SqlServerValueGenerationStrategy.IdentityColumn);
 
                     b.Property<int>("Data")
@@ -38,7 +38,7 @@ namespace L2CacheTest.Migrations
                     b.HasData(
                         new
                         {
-                            Id = 1L,
+                            Id = 1,
                             Data = 1245,
                             Name = "Test"
                         });
